refactor(db): split feed statistics into feed-stats module (#402)
